Renamed root package, decreased min API version, applied Semantic versioning
