+++ app/src/main/AndroidManifest.xml
@@ -1,4 +1,4 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="by.praviloffevg.cryptolib">
+          package="by.praviloffevg.cryptoutils">
 
 </manifest>
