+++ app/src/main/java/by/praviloffevg/cryptoutils/aes/ByteKeyGenerator.kt
@@ -1,4 +1,4 @@
-package by.praviloffevg.cryptolib.aes
+package by.praviloffevg.cryptoutils.aes
 
 import android.util.Log
 import javax.crypto.SecretKeyFactory
