+++ app/src/main/java/by/praviloffevg/cryptoutils/rsa/KeyDurationPeriod.kt
@@ -1,4 +1,4 @@
-package by.praviloffevg.cryptolib.rsa
+package by.praviloffevg.cryptoutils.rsa
 
 import java.util.*
