+++ app/src/main/java/by/praviloffevg/cryptoutils/rsa/KeyProperties.kt
@@ -1,4 +1,4 @@
-package by.praviloffevg.cryptolib.rsa
+package by.praviloffevg.cryptoutils.rsa
 
 data class KeyProperties (
     val keyAlias: String,
