+++ app/src/main/java/by/praviloffevg/cryptoutils/rsa/KeyValidationException.kt
@@ -1,4 +1,4 @@
-package by.praviloffevg.cryptolib.rsa
+package by.praviloffevg.cryptoutils.rsa
 
 class KeyValidationException(
     val code: ExceptionCode,
