+++ app/src/main/java/by/praviloffevg/cryptoutils/rsa/KeyValidationProperties.kt
@@ -1,4 +1,4 @@
-package by.praviloffevg.cryptolib.rsa
+package by.praviloffevg.cryptoutils.rsa
 
 data class KeyValidationProperties(
     val keyDurationPeriod: KeyDurationPeriod,
