+++ app/src/main/java/by/praviloffevg/cryptoutils/rsa/RsaProvider.kt
@@ -1,4 +1,4 @@
-package by.praviloffevg.cryptolib.rsa
+package by.praviloffevg.cryptoutils.rsa
 
 import android.content.Context
 import android.os.Build
