+++ app/src/test/java/by/praviloffevg/cryptoutils/hash/HashUtilTest.kt
@@ -1,9 +1,7 @@
-package by.praviloffevg.cryptolib.hash
+package by.praviloffevg.cryptoutils.hash
 
 import assertk.assert
 import assertk.assertions.isEqualTo
-import by.praviloffevg.cryptolib.hash.Algorithms
-import by.praviloffevg.cryptolib.hash.HashUtil
 import junitparams.JUnitParamsRunner
 import junitparams.Parameters
 import org.junit.Test
